requirements.txt
discord.py[voice]==2.7.1
google-api-python-client==2.192.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.3.0
isort==8.0.1
Jinja2==3.1.6
mdformat==0.7.17
parameterized==0.9.0
playwright==1.58.0
pylint==4.0.5
RapidFuzz==3.14.3